feat: refactor Tabs component to accept tabs and selectedTab as props, removing variant logic
