+++ frontend/src/contexts/MaskingTabContext/index.tsx
@@ -1,3 +1,1 @@
-export { MaskingTabContext, type MaskingTabContextType } from './MaskingTabContext'
 export { MaskingTabProvider } from './MaskingTabProvider'
-export { useMaskingTab } from './useMaskingTab'
